Refactor admin dashboard into modular components
